+++ shared/src/statblock-helpers.ts
@@ -1,5 +1,5 @@
 import YAML from "yaml";
-import type { AttackModel, CasterSpells, FeatureEntity, InnateSpells, SenseEntity, SkillsEntity, SpeedEntity, Stat, Statblock } from "./types";
+import type { AttackModel, CasterSpells, FeatureEntity, InnateSpells, SenseEntity, SkillsEntity, SpeedEntity, SpellSlotList, Stat, Statblock } from "./types";
 
 export const SKILLS_BY_STAT = {
 	str: ["athletics", "strength"],
@@ -144,6 +144,85 @@ export function displaySpeedOrSenses(data: SenseEntity[] | SpeedEntity[], hasEnd
 			output += ", ";
 	}
 	return output.trim();
+}
+
+export function getSpellSlots(sClass: string | null, level: number | null): SpellSlotList | undefined {
+	if (!sClass || !level)
+		return {};
+	if (sClass === "Warlock") {
+		return {
+			1: { 1: 1 },
+			2: { 1: 2 },
+			3: { 2: 2 },
+			4: { 2: 2 },
+			5: { 3: 2 },
+			6: { 3: 2 },
+			7: { 4: 2 },
+			8: { 4: 2 },
+			9: { 5: 2 },
+			10: { 5: 2 },
+			11: { 5: 3 },
+			12: { 5: 3 },
+			13: { 5: 3 },
+			14: { 5: 3 },
+			15: { 5: 3 },
+			16: { 5: 3 },
+			17: { 5: 4 },
+			18: { 5: 4 },
+			19: { 5: 4 },
+			20: { 5: 4 },
+		}[level];
+	}
+	else if (["Ranger", "Paladin", "Artificer"].includes(sClass)) {
+		if (level === 1 && sClass === "Artificer")
+			return { 1: 2 };
+		return {
+			1: {},
+			2: { 1: 2 },
+			3: { 1: 3 },
+			4: { 1: 3 },
+			5: { 1: 4, 2: 2 },
+			6: { 1: 4, 2: 2 },
+			7: { 1: 4, 2: 3 },
+			8: { 1: 4, 2: 3 },
+			9: { 1: 4, 2: 3, 3: 2 },
+			10: { 1: 4, 2: 3, 3: 2 },
+			11: { 1: 4, 2: 3, 3: 3 },
+			12: { 1: 4, 2: 3, 3: 3 },
+			13: { 1: 4, 2: 3, 3: 3, 4: 1 },
+			14: { 1: 4, 2: 3, 3: 3, 4: 1 },
+			15: { 1: 4, 2: 3, 3: 3, 4: 2 },
+			16: { 1: 4, 2: 3, 3: 3, 4: 2 },
+			17: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
+			18: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
+			19: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
+			20: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
+		}[level];
+	}
+	else {
+		return {
+			1: { 1: 2 },
+			2: { 1: 3 },
+			3: { 1: 4, 2: 2 },
+			4: { 1: 4, 2: 3 },
+			5: { 1: 4, 2: 3, 3: 2 },
+			6: { 1: 4, 2: 3, 3: 3 },
+			7: { 1: 4, 2: 3, 3: 3, 4: 1 },
+			8: { 1: 4, 2: 3, 3: 3, 4: 2 },
+			9: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
+			10: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
+			11: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1 },
+			12: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1 },
+			13: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1 },
+			14: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1 },
+			15: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1 },
+			16: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1 },
+			17: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1, 9: 1 },
+			18: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 1, 7: 1, 8: 1, 9: 1 },
+			19: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 2, 7: 1, 8: 1, 9: 1 },
+			20: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 2, 7: 2, 8: 1, 9: 1 },
+		}[level];
+	}
 }
 
 export function spellDc(innate = false, data: Statblock) {